+++ src/packages/property-editors/collection-view/config/bulk-action-permissions/permissions.stories.ts
