+++ src/packages/property-editors/collection-view/config/column/column-configuration.element.ts
@@ -1,4 +1,4 @@
-import type { UmbCollectionColumnConfiguration } from '../../../../../collection/types.js';
+import type { UmbCollectionColumnConfiguration } from '../../../../core/collection/types.js';
 import { html, customElement, property, repeat, css, state, nothing } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import { UmbPropertyValueChangeEvent } from '@umbraco-cms/backoffice/property-editor';
