+++ src/packages/property-editors/collection-view/config/column/column-configuration.test.ts
