+++ src/packages/property-editors/collection-view/config/column/manifests.ts
