+++ src/packages/property-editors/collection-view/config/layout/layout-configuration.element.ts
