+++ src/packages/property-editors/collection-view/config/layout/layout-configuration.test.ts
