+++ src/packages/property-editors/collection-view/config/layout/manifests.ts
