+++ src/packages/property-editors/collection-view/config/order-by/manifests.ts
