+++ src/packages/property-editors/collection-view/manifests.ts
