+++ src/packages/property-editors/collection-view/property-editor-ui-collection-view.element.ts
@@ -1,4 +1,4 @@
-import type { UmbPropertyEditorConfigCollection } from '../../config/index.js';
+import type { UmbPropertyEditorConfigCollection } from '../../core/property-editor/config/index.js';
 import { html, customElement, property, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import { UMB_DOCUMENT_COLLECTION_ALIAS } from '@umbraco-cms/backoffice/document';
