+++ src/packages/property-editors/collection-view/property-editor-ui-collection-view.test.ts
